pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = ["tests", "dev", ".trunk", ".github", ".venv", "dist"]

[tool.hatch.build.targets.wheel]
packages = ["src/google_auth_wrapper"]


[project]
name = "google-auth-wrapper"
version = "0.1.0"
description = "A wrapper for google-auth"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11.7"
dependencies = ["httpx-oauth>=0.10.0", "streamlit>=1.3.0"]

[project.optional-dependencies]
dev = [
  "bandit>=1.8.3",
  "black>=25.1.0",
  "google-cloud-bigquery>=3.30.0",
  "isort>=6.0.1",
  "pytest>=8.3.5",
  "python-dotenv>=1.0.1",
  "ruff>=0.9.9",
]
